get_expert_counts returns Vec<u32> (fixed-width, Uint32Array-safe)
Adopts the one unique improvement from Copilot PR #61: the wasm export
is now fixed-width u32, which keeps the JS dashboard on Uint32Array
(unlike u64/BigInt). State field stays u64 with an export cast.

+++ crates/uor-r4-router/src/lib.rs
@@ -577,10 +577,10 @@ impl UorR4Router {
     }
 
     /// Returns the active counts for the 64 experts
-    pub fn get_expert_counts(&self) -> Vec<usize> {
+    pub fn get_expert_counts(&self) -> Vec<u32> {
         self.expert_active_counts
             .iter()
-            .map(|&c| c as usize)
+            .map(|&c| c as u32)
             .collect()
     }
